+++ resolwe_bio/models.py
@@ -9,7 +9,8 @@ from __future__ import absolute_import, division, print_function, unicode_litera
 
 from django.db import models
 
-from resolwe.flow.models import BaseCollection, Collection
+from resolwe.flow.models import Collection
+from resolwe.flow.models.collection import BaseCollection
 
 
 class Sample(BaseCollection):